Update Vulnerabiltiy Scanner [Bug Fix🐞]
Issue in Vulnerability Scanner: Return statements ends the looping function which attempts to find a banner for each port specified.
Solution:  Adding the results to an array, then at the end of the method return the array; the program can still record the output or lack off and return them for further use.
If this wasn't fixed, no matter how many ports were entered only the first one would run.